Fast_API/requirements.txt
fastapi
uvicorn[standard]
gunicorn
google-cloud-storage